refactor: StaticMeshRenderPass 파일 이름 변경
